requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
behave==1.2.6
    # via -r requirements.in
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   kubernetes
    #   requests
charset-normalizer==3.3.0
    # via requests
google-auth==2.23.2
    # via kubernetes
idna==3.4
    # via requests
kubernetes==28.1.0
    # via -r requirements.in
mysql-connector-python==8.1.0
    # via -r requirements.in
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
parse==1.19.1
    # via
    #   behave
    #   parse-type
parse-type==0.6.2
    # via behave
protobuf==4.21.12
    # via mysql-connector-python
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
python-dateutil==2.8.2
    # via kubernetes
pyyaml==6.0.1
    # via kubernetes
requests==2.31.0
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   behave
    #   kubernetes
    #   parse-type
    #   python-dateutil
urllib3==1.26.17
    # via
    #   kubernetes
    #   requests
websocket-client==1.6.4
    # via kubernetes
